requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pypdf
python-docx
speechrecognition
pyttsx3
deep-translator
plotly
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
torch
numpy
pandas
pyaudio
fastapi
uvicorn
python-multipart
python-dotenv
pydub
